Add a dash to Journal Details column name
When we create this column, we should use dash, underscore, or CamelCase conventions when creating this column name to reflect what we teach learners about about in the Tidy Data lesson. This is something I already do when I teach the lesson, but it should probably be an actual part of the lesson.